move unit tests mocks to unit tests directory
